Refactor: commonboardRepository, commonboardService로 통합
